# api/swagger.json
{"openapi":"3.0.0","paths":{"/auth/userinfo":{"get":{"operationId":"AuthController_getUserInfo","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/lobby/test":{"get":{"operationId":"LobbyController_test","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Lobby"]}},"/lobby/{lobbyId}":{"get":{"operationId":"LobbyController_getLobby","parameters":[{"name":"lobbyId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LobbyResponse"}}}}},"tags":["Lobby"]}},"/lobby/invite":{"post":{"operationId":"LobbyController_generateInvite","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewInviteCodeRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewInviteCodeResponseDto"}}}}},"tags":["Lobby"]}},"/lobby/invite/validate":{"post":{"operationId":"LobbyController_validateInvite","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateInviteCodeRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateInviteCodeResponseDto"}}}}},"tags":["Lobby"]}},"/lobby/validateCreator":{"post":{"operationId":"LobbyController_validateCreator","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateCreatorSecretRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidateCreatorSecretResponseDto"}}}}},"tags":["Lobby"]}},"/lobby":{"post":{"operationId":"LobbyController_postLobby","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLobbyRequest"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LobbyResponse"}}}}},"tags":["Lobby"]}},"/lobby/increment":{"post":{"operationId":"LobbyController_addPointsToLobby","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddPixelsRequest"}}}},"responses":{"201":{"description":""}},"tags":["Lobby"]},"patch":{"operationId":"LobbyController_editLobbyPoints","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditPixelsRequest"}}}},"responses":{"200":{"description":""}},"tags":["Lobby"]}},"/lobby/increment/confirm":{"patch":{"operationId":"LobbyController_confirmIncrement","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmIncrementRequest"}}}},"responses":{"200":{"description":""}},"tags":["Lobby"]}},"/lobby/accept/{lobbyId}/{code}":{"get":{"operationId":"LobbyController_acceptInvite","parameters":[{"name":"lobbyId","required":true,"in":"path","schema":{"type":"string"}},{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Lobby"]}},"/lobby/reject/{lobbyId}/{code}":{"get":{"operationId":"LobbyController_rejectInvite","parameters":[{"name":"lobbyId","required":true,"in":"path","schema":{"type":"string"}},{"name":"code","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["Lobby"]}},"/lobby/my-lobbies":{"post":{"operationId":"LobbyController_myLobbies","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LobbyPreviewResponse"}}}}}},"tags":["Lobby"]}}},"info":{"title":"PaintIncrementApi","description":"The API definition for Paint Increment","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"IncrementPixel":{"type":"object","properties":{"x":{"type":"number"},"y":{"type":"number"}},"required":["x","y"]},"IterationModel":{"type":"object","properties":{"name":{"type":"string"},"id":{"type":"string"},"pixels":{"type":"array","items":{"$ref":"#/components/schemas/IncrementPixel"}},"confirmed":{"type":"boolean"}},"required":["name","id","pixels","confirmed"]},"PaintLobbySettings":{"type":"object","properties":{"width":{"type":"number"},"height":{"type":"number"},"maxPixels":{"type":"number"},"timeLimit":{"type":"number"}},"required":["width","height","maxPixels","timeLimit"]},"LobbyResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"pixelIterations":{"type":"array","items":{"$ref":"#/components/schemas/IterationModel"}},"settings":{"$ref":"#/components/schemas/PaintLobbySettings"},"isCreator":{"type":"boolean"}},"required":["id","name","pixelIterations","settings","isCreator"]},"NewInviteCodeRequestDto":{"type":"object","properties":{"lobbyId":{"type":"string"}},"required":["lobbyId"]},"NewInviteCodeResponseDto":{"type":"object","properties":{"inviteCode":{"type":"string"}},"required":["inviteCode"]},"ValidateInviteCodeRequestDto":{"type":"object","properties":{"lobbyId":{"type":"string"},"inviteCode":{"type":"string"}},"required":["lobbyId","inviteCode"]},"ValidateInviteCodeResponseDto":{"type":"object","properties":{"isValid":{"type":"boolean"}},"required":["isValid"]},"ValidateCreatorSecretRequestDto":{"type":"object","properties":{"lobbyId":{"type":"string"}},"required":["lobbyId"]},"ValidateCreatorSecretResponseDto":{"type":"object","properties":{"isValid":{"type":"boolean"}},"required":["isValid"]},"CreateLobbyRequest":{"type":"object","properties":{"name":{"type":"string"},"ownerName":{"type":"string"},"settings":{"$ref":"#/components/schemas/PaintLobbySettings"}},"required":["name","ownerName","settings"]},"AddPixelsRequest":{"type":"object","properties":{"lobbyId":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"pixels":{"type":"array","items":{"$ref":"#/components/schemas/IncrementPixel"}},"inviteCode":{"type":"string"}},"required":["lobbyId","name","email","pixels"]},"EditPixelsRequest":{"type":"object","properties":{"lobbyId":{"type":"string"},"pixels":{"type":"array","items":{"$ref":"#/components/schemas/IncrementPixel"}},"incrementId":{"type":"string"}},"required":["lobbyId","pixels","incrementId"]},"ConfirmIncrementRequest":{"type":"object","properties":{"lobbyId":{"type":"string"},"accept":{"type":"boolean"}},"required":["lobbyId","accept"]},"LobbyPreviewResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"}},"required":["id","name"]}}}}